Slår sammen innsending skjemaene - wip
